bug fix
Issue: If twitter responds with 0 or more tweets and has_more_items is false, but Since date has not reached. This means the query is not complete but program is thinking it is, based on Twitter response.
Now program will try to loop through 5 times even if has_more_items is false. Then if there are still no tweets, then program will reste until date to either last tweet date or current until date minus 1 day (-1). and try again. This way program will keepsearching until the since date is reached.